Move wire_doppelganger_pool to header
