+++ shards/modules/core/wires.cpp
@@ -9,6 +9,7 @@
 #include <shards/common_types.hpp>
 #include <shards/core/foundation.hpp>
 #include <shards/core/ops_internal.hpp>
+#include <shards/core/wire_doppelganger_pool.hpp>
 #include <shards/shards.h>
 #include <shards/shards.hpp>
 #include <chrono>
