+++ shards/modules/network/network.cpp
@@ -8,6 +8,7 @@
 #include <shards/shards.hpp>
 #include <shards/shardwrapper.hpp>
 #include <shards/core/shared.hpp>
+#include <shards/core/wire_doppelganger_pool.hpp>
 #include <shards/utility.hpp>
 #include <optional>
 #include <boost/lockfree/queue.hpp>
